+++ blocklets/ai-kit/src/pages/playground/chat.tsx
@@ -1,12 +1,55 @@
+import Conversation, {
+  ConversationRef,
+  MessageItem,
+  useConversation,
+} from '@blocklet/ai-kit/src/components/conversation/index';
 import Dashboard from '@blocklet/ui-react/lib/Dashboard';
 import styled from '@emotion/styled';
+import { HighlightOff } from '@mui/icons-material';
+import { Button, Tooltip } from '@mui/material';
+import { ReactNode, useCallback, useRef } from 'react';
 
-import Conversation from '../../components/conversation';
+import { ImageGenerationSize, imageGenerations, textCompletions } from '../../libs/ai';
 
 export default function Chat() {
+  const ref = useRef<ConversationRef>(null);
+
+  const { messages, add, cancel } = useConversation({
+    scrollToBottom: (o) => ref.current?.scrollToBottom(o),
+    textCompletions: (prompt) => textCompletions({ prompt, stream: true }),
+    imageGenerations: (prompt) =>
+      imageGenerations({ ...prompt, size: prompt.size as ImageGenerationSize, response_format: 'b64_json' }).then(
+        (res) => res.data.map((i) => ({ url: `data:image/png;base64,${i.b64_json}` }))
+      ),
+  });
+
+  const customActions = useCallback(
+    (msg: MessageItem): [ReactNode[], ReactNode[]] => {
+      return [
+        [],
+        [
+          msg.loading && (
+            <Tooltip key="stop" title="Stop" placement="top">
+              <Button size="small" onClick={() => cancel(msg)}>
+                <HighlightOff fontSize="small" />
+              </Button>
+            </Tooltip>
+          ),
+        ],
+      ];
+    },
+    [cancel]
+  );
+
   return (
     <Root footerProps={{ className: 'dashboard-footer' }}>
-      <Conversation maxWidth={800} />
+      <Conversation
+        ref={ref}
+        sx={{ maxWidth: 800, mx: 'auto', width: '100%' }}
+        messages={messages}
+        onSubmit={(prompt) => add(prompt)}
+        customActions={customActions}
+      />
     </Root>
   );
 }
@@ -24,6 +67,12 @@ const Root = styled(Dashboard)`
     > .dashboard-footer {
       margin-top: 0;
       padding: 0;
+
+      .logo-container {
+        svg {
+          height: 100%;
+        }
+      }
     }
   }
 `;