+++ packages/react/src/components/conversation/index.ts
@@ -1,0 +1,6 @@
+import Conversation from './conversation';
+
+export default Conversation;
+export * from './conversation';
+
+export { default as useConversation } from './use-conversation';