+++ packages/react/src/index.tsx
@@ -1,5 +1,6 @@
 import ImagePreview from './components/image-preview';
 import LoadingImage from './components/loading-image';
 
-// eslint-disable-next-line import/prefer-default-export
 export { ImagePreview, LoadingImage };
+
+export * from './api/ai-kit';